added pause and resume functionality
There is now a 'Pause'-Method and a 'Resume'-Method on every node.
When Pause() is called on the behavior tree (the root), the current active tasks are stopped
but without affecting the decorators and composites stop logic.
When Resume() is called the previously stopped tasks are started again.

Pause() and Resume() should't be called on other nodes but the root.

Instead of changing the 'Root', 'Task' and 'Composite' class as suggested, I just changed the common parent class 'Container'.