instead of checking if the container is active/paused and fail silently, asserting the right values
Because Pause() is only cascaded on active nodes, we can assume that Pause() is only called on active node. The only exception is the first node on which Pause() is called, therefore there is an assert checking this condition.

There is also an assert checking that Resume() is only called on a paused container.