Add check that lower priority nodes aren't stopped in pause state (also no restart)
When a ObservingDecorator is Paused, the decorator stops observing.

But there is also the case where another child in the parent composite is paused and the ObservingDecorator is inactive but still observes for changes to do his logic depending on the stops behavior.
For this case there is now an additional check so that no evaluation for stopping and restarting happens in the pause state.