Fix problems with blackboard condition not evaluated after resuming
To fix the problem, I changed that an inactive blackboard condition within a paused composite is actually also paused, so that after resuming an evaluation happens.

For this to happen I changed the Pause() method in Container so that Pause() is always called even if the container is active. With exceptions to the tasks, on them Pause() should only be called when there are active, so that they can be stopped.
Then the child is only added to the pausedChildren when the state is paused or when it is a stopped task.
In ObservingDecorator the method Pause() was then overwritten to always
pause, no matter in which state in, but only propagate Pause() on the children when active.
Because it's Paused it will be evaluated after Resume().

The only other tricky thing was the StopLowerPriorityChildrenForChild() method in Composite. This method searches for a lower priority active child and thus it's important when resuming that the lower priority nodes are resumed before the higher priority nodes (from right to left).
So we need to resume from right to left. Therefore a Stack for the pausedChildren was used so that the nodes are resumed from right to left.